renamed loss_plot.py to loss.py
