+++ src/careamics/plotting/__init__.py
@@ -1,6 +1,6 @@
 """Plotting Module."""
 
-from .loss_plot import plot_loss
+from .loss import plot_loss
 from .noise_model import plot_noise_model_distribution
 from .noise_residual import plot_noise_residuals
